+++ CHANGELOG.md
@@ -9,6 +9,7 @@ v1.16.0-dev (unreleased)
 - transport/http: Add GrabHeaders option to propagate specific
   headers starting with x- to handlers.
 
+
 v1.15.0 (2017-09-15)
 --------------------
 
@@ -18,6 +19,7 @@ v1.15.0 (2017-09-15)
   based on the transport encoding.
 - transport/x/grpc: Add options to specify the maximum message size
   sent and received over the wire.
+
 
 v1.14.0 (2017-09-08)
 --------------------